requirements.txt
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
python-dotenv==1.0.1
pandas==2.0.0
requests==2.31.0
alembic==1.11.1
psycopg2-binary==2.9.10